lift out url.ast.Name into ermpath to support wildcard expansion
